tools/requirements.txt
requests
python-dotenv
elasticsearch